+++ src/__tests__/personProcessing.test.ts
@@ -33,7 +33,7 @@ jest.mock('../utils/globals', () => {
     }
 })
 
-// eslint-disable-next-line @typescript-eslint/no-var-requires
+// eslint-disable-next-line @typescript-eslint/no-require-imports
 const { mockURLGetter, mockReferrerGetter } = require('../utils/globals')
 
 describe('person processing', () => {